Refactor project structure, update API calls, and improve server startup logic
