+++ src/main.jsx
@@ -4,7 +4,7 @@ import { createRoot } from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import App from './App.jsx';
 import './index.css';
-import { AuthProvider } from './context/AuthContext.jsx'; // Corrected import path
+import { AuthProvider } from './server/AuthContext.jsx';
 
 const root = createRoot(document.getElementById('root'));
